pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-bootstrap5==25.1",
    "django-filter>=25.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.2",
]

[tool.ruff]
line-length = 79
select = ["E", "W", "F", "I"]
ignore = ["E501"]
extend-exclude = ["venv", ".venv"]